up: move some request and vo dto to subdir.
- remove some not used dto class
