+++ README.md
@@ -1,11 +1,14 @@
-# openjob
+# OpenJob
+
 Distributed task scheduling
 
-# swagger
+## Develop
+
+### Swagger
 
 http://127.0.0.1:8080/swagger-ui.html
 
-# openjob-ui
+### Openjob-ui
 
 ```shell
 git submodule update --init  //初始化版本
@@ -14,9 +17,14 @@ git submodule update --remote  //更新到最新版本
 cd openjob-ui & git pull
 ```
 
-# repository
+### Repository
+
 - save
 - get
 - list
 - update
 - delete
+
+## License
+
+Please see [LICENSE](LICENSE)
