+++ openjob-server/openjob-server-admin/src/main/java/io/openjob/server/admin/request/menu/AdminMenuDeleteRequest.java
@@ -1,4 +1,4 @@
-package io.openjob.server.admin.request;
+package io.openjob.server.admin.request.menu;
 
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
@@ -11,12 +11,12 @@ import javax.validation.constraints.NotNull;
 
 /**
  * @author inhere
- * @date 2022-11-13 23:27:29
+ * @date 2022-11-13 23:27:31
  * @since 1.0.0
  */
 @Data
-@ApiModel(value = "AdminMenuQuery", description = "Query AdminMenu")
-public class AdminMenuQueryRequest {
+@ApiModel(description = "Delete AdminMenu")
+public class AdminMenuDeleteRequest {
 
     @Min(1)
     @NotNull()
