+++ openjob-server/openjob-server-admin/src/main/java/io/openjob/server/admin/service/impl/AdminMenuServiceImpl.java
@@ -3,14 +3,14 @@ package io.openjob.server.admin.service.impl;
 import io.openjob.server.common.constant.CommonConst;
 import io.openjob.server.repository.data.AdminMenuData;
 import io.openjob.server.admin.service.AdminMenuService;
-import io.openjob.server.admin.request.AdminMenuAddRequest;
-import io.openjob.server.admin.vo.AdminMenuAddVO;
-import io.openjob.server.admin.request.AdminMenuQueryRequest;
-import io.openjob.server.admin.vo.AdminMenuQueryVO;
-import io.openjob.server.admin.request.AdminMenuUpdateRequest;
-import io.openjob.server.admin.vo.AdminMenuUpdateVO;
-import io.openjob.server.admin.request.AdminMenuDeleteRequest;
-import io.openjob.server.admin.request.AdminMenuListRequest;
+import io.openjob.server.admin.request.menu.AdminMenuAddRequest;
+import io.openjob.server.admin.vo.menu.AdminMenuAddVO;
+import io.openjob.server.admin.request.menu.AdminMenuQueryRequest;
+import io.openjob.server.admin.vo.menu.AdminMenuQueryVO;
+import io.openjob.server.admin.request.menu.AdminMenuUpdateRequest;
+import io.openjob.server.admin.vo.menu.AdminMenuUpdateVO;
+import io.openjob.server.admin.request.menu.AdminMenuDeleteRequest;
+import io.openjob.server.admin.request.menu.AdminMenuListRequest;
 import io.openjob.server.repository.dto.AdminMenuDTO;
 import io.openjob.server.common.dto.PageDTO;
 import org.springframework.beans.BeanUtils;
