+++ openjob-server/openjob-server-admin/src/main/java/io/openjob/server/admin/vo/menu/AdminMenuQueryVO.java
@@ -1,23 +1,20 @@
-package io.openjob.server.admin.request;
+package io.openjob.server.admin.vo.menu;
 
-import io.openjob.server.admin.request.part.MenuMeta;
+import io.openjob.server.admin.vo.part.MenuMeta;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
 import java.util.List;
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
 
 /**
  * @author inhere
- * @date 2022-11-13 23:27:28
+ * @date 2022-11-13 23:24:42
  * @since 1.0.0
  */
 @Data
-@ApiModel(value = "AdminMenuUpdate", description = "Update AdminMenu")
-public class AdminMenuUpdateRequest {
+@ApiModel(value = "AdminMenu", description = "AdminMenu")
+public class AdminMenuQueryVO {
 
     @ApiModelProperty(value = "PK")
     private Integer id;
@@ -46,5 +43,13 @@ public class AdminMenuUpdateRequest {
     @ApiModelProperty(value = "Delete status. 1=yes 2=no")
     private Integer deleted;
 
+    @ApiModelProperty(value = "Delete time")
+    private Integer deleteTime;
+
+    @ApiModelProperty(value = "Update time")
+    private Integer updateTime;
+
+    @ApiModelProperty(value = "Create time")
+    private Integer createTime;
 }
