+++ openjob-server/openjob-server-admin/src/main/java/io/openjob/server/admin/vo/menu/AdminMenuUpdateVO.java
@@ -1,4 +1,4 @@
-package io.openjob.server.admin.vo;
+package io.openjob.server.admin.vo.menu;
 
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
